+++ jest.setup.ts
@@ -1,0 +1,5 @@
+// For some reason sometimes this setup is not running (yielding errors for Zone being
+// undefined). This is here to help in diagnosing the issue.
+
+import "reflect-metadata";
+import "zone.js";