+++ src/rpc-proxy.test.ts
@@ -1,11 +1,11 @@
 import { expect } from "chai";
-import { describe } from "razmin";
+import { describe, it } from "@jest/globals";
 import { RPCProxy } from "./rpc-proxy";
 import { RPCSession } from "./session";
 import { TestChannel } from "./test-channel";
 import { getRpcType } from "./internal";
 
-describe('RPCProxy', it => {
+describe('RPCProxy', () => {
     it('is remotable', () => {
         let [a, b] = TestChannel.makePair();
         let sessionA = new RPCSession(a);