+++ src/test.ts
@@ -1,4 +1,0 @@
-import "reflect-metadata";
-import { suite } from 'razmin';
-
-suite().include(['**/*.test.js']).run();